Honour the "--part" command line argument
Bisection shows that the argument is ignored since commit e095b38d,
the porting of KonqRun to UrlLoader.  If set, use it to find the
service for the part in UrlLoader::decideEmbedOrSave().

Test cases:
without patch konqueror --part khtml kde.org --> displays in WebEngine
with this patch --> displays in KHTML