Add back the View Mode menu
Since KPluginMetaData provides no replacement for KService::actions,
implementing this menu for parts providing more than one view mode (in
particular, dolphinpart) is problematic. For now, we assume that such
parts provide a separate .desktop file and that its name is stored in an
"actionFile" entry of the plugin metadata. This file must have all the
action-related content of the old part .desktop file, organized in the
same way.

TODO: currently, it's assumed that the .desktop file containing the
actions is stored in Qt::StandardPaths::DataLocation and that the
"actionFile" entry in the plugin metadata contains the path of the file
relative to that directory. Find out whether this is correct