Emit the openUrlRequest from a QTimer::singleShot
Middle-clicking on a link in a WebEnginePage correctly opens the link in
a new tab, but, incorrectly, leaves the focus in the location bar
instead of giving it to the new widget. Using a QTimer::singleShot to
ensure that the request to open the URL is made after all events have
been processed prevents this issue.

Note that this is only a workaround: it doesn't fix the true cause of
the problem.

1f014821e987d52b7e9574db7153dc4c2b147a34


(cherry picked from commit 78f47b4d4ce97f7ff0278b16c04182e7d80c8dba)