Fix crash caused by using original view rather than actual view
UrlLoader::decideEmbedOrSave used the original view when deciding
whether it could be reused to open the URL, instead of using the view
provided using UrlLoader::setView. If the URL was opened by clicking on
a sidebar item, this caused a sidebar view to be used, rather than the
correct dolphin part. This lead to a crash because the sidebar view is
a passive view, which caused KonqViewManager::chooseNextView to return
nullptr

BUG: 458933


(cherry picked from commit c9f3d1957f7b6c168a025b5b4014e46a5f49bb63)