Improve choice of window when a new tab is requested by kfmclient
Previously, the new tab was always created in the oldest window in the
current desktop. Now, instead, the tab is created in the window which
was last used. Besides, the window must not only be in the current desktop,
but also in the current activity.


(cherry picked from commit bc3685ef4e856e9311533b12a72660f73bb9251d)