Use WebEnginePart to download URLs before opening them
This is in preparation for the removal of kcookiejar in KF6. Since
kcookiejar will be removed, there won't be any way for WebEnginePart to
share cookies with applications which use KIO to load the URLs to open.
This would create problems to open which can only be accessed after
authentication. To avoid this problem, now WebEnginePart download the
URL to a local file and the appropriate part or application is then
asked to open the local file.

NOTE: this is only enabled when Konqueror is build with Qt6