Use custom profile class when building for KF6
In Qt6 QWebEngineProfile::defaultProfile() returns an off-the-record
profile, which is unsuitable for common use. Rather than implementing
something like WebEnginePart::defaultProfile, a custom profile class is
introduced, providing a defaultProfile() static function to return a
suitable default profile. This way, later we'll be able to move the
functionality of WebEnginePartControls in the custom profile and remove
WebEnginePartControls.