Dowonload URL in global temporary directory when opening it in external application
When the requesting part wants to download the URL itself and the URL
should then be opened in an external application, downloading the URL in
a Konqueror-specific temporary directory (/tmp/WebEnginePartDownloadManager*
for example) could lead to problems if Konqueror was closed before the
external application, since the directory, and the file itself, would be
deleted.

To avoid this, in this situation the URL is downloaded in the global
temporary directory (/tmp/) instead. The downside is that, if the
external application doesn't support the --tmpfile switch, the file
won't be automatically removed.