Allow Konqueror as a file manager in the "Default Applications" chooser
For an application to be shown in the list for "File Manager", it
must support the inode/directory MIME type and have "FileManager"
included in its application categories.  The NoDisplay option is
ignored there.

See plasma-desktop/kcms/componentchooser/componentchooserfilemanager.cpp
and plasma-desktop/kcms/componentchooser/componentchooser.cpp
for the code that implements the application trader query.