Save and read cookies on disk
It seems that after calling QWebEngineCookieStore::loadCookies the
cookieAdded signal isn't emitted until you visit a web page and then
only for the cookies related to that page. This makes the cookie
management tab not work correctly. To work around this issue, save the
cookies in a file on destruction and load them back on creation.