Check if the value passed to runJavascript callbacks are valid
The callback can be called during the destruction phase when, according
to documentation, it isn't safe to access the page or the view. When
this happens, the variant passed as argument to the callback is invalid.
Having the callback do nothing when this happens prevents crashes.


(cherry picked from commit c7231073673534a23949741deeabcc40975544bc)