Fix a bug causing Konqueror to hang when filling a bookmark menu with many bookmarks
The cause of the bug is that if the menu contains many bookmarks and the
favicon cache is empty, a large number of KIO::FavIconRequestJob will be
created at the same time. Moreover, a changed signal will be emitted
after each request, triggering the connected slot for each action.

To avoid this, the number of running jobs is capped and a new function
which emits the changed signal only once is introduced.