UrlLoader: Fix resolving of embedding part to use
Rearrange the order of the first two checks so that, if a part is
explicitly specified via the "--part" command line option, it is
used as the first priority.

The original first test ("if m_view && ...") would always be true
on first startup because the view exists at this point, showing
about:blank via the WebEnginePart.  This part accepts most textual
MIME types and would claim the URL, ignoring the command line option.