Add option to always embed files from web pages in a new tab
Some web pages don't support navigating back using the browser "back" button. If on one such page you click on a link which causes a file to be embedded in the same view, you are unable to go back to the original page. In some cases, if the user knows of this problem, he can middle click on the link to open it in a new tab. However, this doesn't work if there's a button instead of a link. This option works around this problem by causing all files from web pages which are to be embedded to be put in a new tab.