i18n: extract strings from sidebar/
For some reasons they were not extracted. One of the items
inside sidebar referenced the konqueror translation domain,
but its string were not extracted anyway.

Define a new translation domain and extract all strings
(but the ones from tree/, which is disabled, not ported even
to Qt4 and which should be removed or ported), and set
the translation domain accordingly.

It would be a bit too much to define a separate translation
file for each sidebar item, and in fact this is the case
for kcmkonqhtml, which defines a single translation domain
for several plugins/libraries.

BUG: 500537
(cherry picked from commit c81cf007d1e883b4ff88dd01e7ad6100be3ec66c)